+++ tests/GdTestCase.php
@@ -14,7 +14,7 @@ abstract class GdTestCase extends BaseTestCase
 {
     public static function readTestImage(string $filename = 'test.jpg'): Image
     {
-        return (new Driver())->specialize(new FilePathImageDecoder())->decode(
+        return (new Driver())->specializeDecoder(new FilePathImageDecoder())->decode(
             static::getTestResourcePath($filename)
         );
     }