+++ tests/ImagickTestCase.php
@@ -16,7 +16,7 @@ abstract class ImagickTestCase extends BaseTestCase
 {
     public static function readTestImage(string $filename = 'test.jpg'): Image
     {
-        return (new Driver())->specialize(new FilePathImageDecoder())->decode(
+        return (new Driver())->specializeDecoder(new FilePathImageDecoder())->decode(
             static::getTestResourcePath($filename)
         );
     }